Adjust auth common component location
