+++ src/components/Graph/LandingPage/SearchBar.tsx
@@ -84,7 +84,7 @@ const Searchbar: FC<SearchbarProps> = ({ className, onSearchAddress }) => {
             }
           }}
           className={
-            "block w-full rounded-none rounded-l-md border-0 py-1.5 pl-10 font-mono text-gray-900 ring-1 ring-inset ring-gray-300 transition-all placeholder:text-gray-400 focus:ring-2 focus:ring-inset sm:text-sm sm:leading-6" +
+            "block w-full rounded-none rounded-l-md border-0 py-1.5 pl-10 font-mono text-sm leading-6 text-gray-900 ring-1 ring-inset ring-gray-300 transition-all placeholder:text-gray-400 focus:ring-2 focus:ring-inset" +
             (isAddressValid || !query
               ? " focus:ring-blue-600"
               : " focus:ring-red-500")
